Set watt/powerlimit with nvidia-overclock.sh
Added so that power limit is set just as every other overclocking feature after tweaking in the setup. Previously it was only set when miner.sh ran, now it's both (which is probably correct). 

Might consider removing watt in miner.sh since it runs on bootup and if boot-state is supposed to be "fail-safe" forcing a watt setting that early might cause issues. Better to have cron run it later, as with other OC.