Environment variables for launch of ethminer
Since version 0.13 the express export of GPU_x environment variables is
no longer needed as they are already embedded into ethminer itself.
On the other hand is relevant to add CUDA_DEVICE_ORDER as this
guarantess that CUDAx kernels are numbered as the order of cards on the
pci bus. This is important for debugging and overclocking.
Furthermore the --farm-recheck parameter has no meaning anymore: in
stratum mode it only affected the display interval of hashrate while
since release 0.14 of ethminer it has been completely dropped (in
stratum) as the hashrate display interval has been fixed to 5 seconds.